Assignment Rails Labels: completed

+++ app/controllers/comments_controller.rb
@@ -3,29 +3,58 @@ class CommentsController < ApplicationController
   before_action :authorize_user, only: [:destroy]
 
   def create
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.new(comment_params)
-    comment.user = current_user
+    if params.include?(:post_id)
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
+      comment.user = current_user
 
-    if comment.save
-      flash[:notice] = "Comment saved successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment failed to save."
-      redirect_to [@post.topic, @post]
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to [@post.topic, @post]
+      end
+
+    elsif params.include?(:topic_id)
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+      comment.user = current_user
+
+      if comment.save
+        flash[:notice] = "Comment saved successfully."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment failed to save."
+        redirect_to @topic
+      end
     end
   end
 
   def destroy
-    @post = Post.find(params[:post_id])
-    comment = @post.comments.find(params[:id])
+    if params.include?(:post_id)
+      @post = Post.find(params[:post_id])
+      comment = @post.comments.new(comment_params)
 
-    if comment.destroy
-      flash[:notice] = "Comment was deleted successfully."
-      redirect_to [@post.topic, @post]
-    else
-      flash[:alert] = "Comment couldn't be deleted. Try again."
-      redirect_to [@post.topic, @post]
+      if comment.destroy
+        flash[:notice] = "Comment deleted successfully."
+        redirect_to [@post.topic, @post]
+      else
+        flash[:alert] = "Comment failed to delete."
+        redirect_to [@post.topic, @post]
+      end
+
+    elsif params.include?(:topic_id)
+      @topic = Topic.find(params[:topic_id])
+      comment = @topic.comments.new(comment_params)
+
+      if comment.destroy
+        flash[:notice] = "Comment deleted successfully."
+        redirect_to @topic
+      else
+        flash[:alert] = "Comment failed to delete."
+        redirect_to @topic
+      end
     end
   end
 
@@ -42,5 +71,5 @@ class CommentsController < ApplicationController
        redirect_to [comment.post.topic, comment.post]
      end
    end
-   
+
 end
